fix: Merged snapshot_reader.go and snapshot_writer.go
